+++ src/pattern/background.worker/worker.interface.ts
@@ -1,0 +1,13 @@
+import {Subject} from 'rxjs';
+
+export interface IWorkerInput<T> {
+    scriptPath: string;
+    data: T;
+}
+
+export interface IBackgroundWorker<T, K> {
+    WorkPath: string;
+    OnFinish: Subject<K>;
+    OnError: Subject<Error>;
+    Run(args?: T);
+}