+++ webpack.config.js
@@ -1,0 +1,35 @@
+const path = require('path');
+
+module.exports = {
+    entry: './src/ts-tooling.ts',
+    mode: 'development',
+    resolve: {
+        extensions: ['.tsx', '.ts', '.js']
+    },
+    module: {
+        rules: [
+            {
+                test: /\.tsx?$/,
+                loader: 'ts-loader?configFile=tsconfig.json',
+                exclude: [/node_modules/]
+            }
+        ]
+    },
+    output: {
+        path: path.resolve(__dirname, 'lib'),
+        filename: 'ts-tooling.js',
+        library: 'tst',
+        libraryTarget: 'umd',
+        globalObject: 'this',
+    },
+    externals: {
+        fs: {
+            commonjs: 'fs',
+            commonjs2: 'fs',
+        },
+        worker_threads: {
+            commonjs: 'worker_threads',
+            commonjs2: 'worker_threads',
+        },
+    },
+};